requirements.txt
anyio==4.0.0
APScheduler==3.10.4
astroid==3.0.1
black==23.11.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dill==0.3.7
dnspython==2.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
isort==5.12.0
logzero==1.7.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pylint==3.0.2
pymongo==4.6.0
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tomlkit==0.12.2
tzlocal==5.2
watchdog==3.0.0
